updated graphic color options

+++ AltairHL_emulator/Drivers/pi_sense_hat/graphics.h
@@ -21,4 +21,4 @@ void gfx_bitmap_to_rgb(uint8_t bitmap[8], uint16_t *panel_buffer, size_t buffer_
 void gfx_load_character(uint8_t character, uint8_t bitmap[8]);
 void gfx_reverse_panel(unsigned char A[8]);
 void gfx_rotate_counterclockwise(unsigned char A[8], uint32_t m, uint32_t n, unsigned char B[8]);
-void gfx_set_color(uint16_t color);
+void gfx_set_color(uint8_t color);
